Rename Export module to Spreadsheet and update CLI
